refactor : ViewModel 공통 기능(BaseViewModel) 추가 및 Content Control 바인딩
- INotifyPropertyChanged 공통 구현을 위해 BaseViewModel 클래스 추가
- MainView에 Content Control 바인딩 구성 (CurrentView)
