+++ QR_Tracker/View/MainView.xaml.cs
@@ -11,17 +11,16 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
+using QR_Tracker.ViewModel;
 
 namespace QR_Tracker.View
 {
-    /// <summary>
-    /// MainView.xaml에 대한 상호 작용 논리
-    /// </summary>
     public partial class MainView : Window
     {
         public MainView()
         {
             InitializeComponent();
+            DataContext = new MainViewModel();
         }
     }
 }
