+++ QR_Tracker/ViewModel/QrDetectViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using QR_Tracker.ViewModel.BaseViewModels;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
@@ -7,8 +8,9 @@ using System.Threading.Tasks;
 
 namespace QR_Tracker.ViewModel
 {
-    public class QrDetectViewModel 
+    public class QrDetectViewModel : BaseViewModel
     {
+        public string TestMessage => "뷰모델이 연결됐어요!";
 
     }
 }
